+++ sql/selling_history.sql
@@ -1,8 +1,8 @@
 CREATE TABLE selling_history(mobilenumber BIGINT NOT NULL,
 							 product_name VARCHAR(50) NOT NULL,
-							 price_per_quantity REAL NOT NULL,
+							 price_per_quantity DECIMAL(5,2) NOT NULL,
 							 ordered_quantity INT NOT NULL,
-							 amount REAL NOT NULL,
+							 amount DECIMAL(5,2) NOT NULL,
 							 buy_date DATE DEFAULT CURRENT_DATE,
 							 buy_time  TIME DEFAULT CURRENT_TIME);
 SELECT * FROM selling_history;
